add embedded redis cache

+++ SpingBootCacheRedis/src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS person;
 create table person
 (
-id integer primary key auto_increment,
+id NUMERIC(19,0) primary key auto_increment,
 name varchar(64),
 age integer,
 address varchar(64)
